src/skyvern/requirements.txt
skyvern
python-dotenv
